+++ src/main/java/cn/edu/seig/novel/dto/resp/BookCommentRespDto.java
@@ -33,7 +33,7 @@ public class BookCommentRespDto {
 
         private String commentUserPhoto;
 
-        private Integer rate;
+        private Float rate;
 
         @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
         private LocalDateTime commentTime;
